feat: Integrate DashboardTop component into Dashboard
